+++ tools/devsecops_engine_tools/engine_sca/engine_dependencies/src/domain/usecases/dependencies_sca_scan.py
@@ -13,12 +13,16 @@ class DependenciesScan:
         tool_deserializator: DeserializatorGateway,
         remote_config,
         dict_args,
+        exclusions,
+        pipeline_name,
         to_scan,
         token,
     ):
         self.tool_run = tool_run
         self.tool_deserializator = tool_deserializator
         self.remote_config = remote_config
+        self.exclusions = exclusions
+        self.pipeline_name = pipeline_name
         self.dict_args = dict_args
         self.to_scan = to_scan
         self.token = token
@@ -32,6 +36,8 @@ class DependenciesScan:
         return self.tool_run.run_tool_dependencies_sca(
             self.remote_config,
             self.dict_args,
+            self.exclusions,
+            self.pipeline_name,
             self.to_scan,
             self.token,
         )
