+++ tools/devsecops_engine_tools/engine_sca/engine_dependencies/test/infrastructure/entry_points/test_entry_point_tool.py
@@ -12,9 +12,7 @@ def test_init_engine_dependencies():
         "devsecops_engine_tools.engine_sca.engine_dependencies.src.infrastructure.entry_points.entry_point_tool.SetInputCore"
     ) as mock_set_input_core, patch(
         "devsecops_engine_tools.engine_sca.engine_dependencies.src.infrastructure.entry_points.entry_point_tool.HandleRemoteConfigPatterns"
-    ) as mock_handle_remote_config_patterns, patch(
-        "devsecops_engine_tools.engine_sca.engine_dependencies.src.infrastructure.entry_points.entry_point_tool.FindArtifacts"
-    ) as mock_find_artifacts:
+    ) as mock_handle_remote_config_patterns:
         dict_args = {"remote_config_repo": "remote_repo"}
         token = "token"
         tool = "tool"
